requirements.txt
ImageHash==4.3.1
numpy==2.0.0
pillow==10.3.0
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
PyWavelets==1.6.0
scipy==1.13.1
shiboken6==6.7.2
